iss-location/requirements.txt
google-adk
python-dotenv
google-cloud-aiplatform
google-auth
fastmcp
requests